# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "UPic API",
        "description": "Upic API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "Authorization"
        }
    },
    "schemes": ["http", "https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/comments": {
            "post": {
                "tags": ["Comments"],
                "description": "Create a new comment",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "The comment that we want to create",
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    }
                ],
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "description": "New comment is created",
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    }
                }
            },
            "get": {
                "tags": ["Comments"],
                "summary": "Get all comments in the system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    },
                    "404": {
                        "description": "Another error",
                        "schema": {
                            "example": "Another Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "example": "Internal server error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "User": {
            "required": ["nom", "email", "_id", "password"],
            "properties": {
                "_id": {
                    "type": "string",
                    "uniqueItems": true
                },
                "email": {
                    "type": "string",
                    "uniqueItems": true
                },
                "lastName": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "uniqueItems": true
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Image"
                    }
                },
                "likedImages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Image"
                    }
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Comment"
                    }
                },
                "abonnes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User"
                    }
                },
                "abonnements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User"
                    }
                },
                "nAbonnes": {
                    "type": "number"
                }
            }
        },
        "Comment": {
            "required": ["_id", "comment", "user", "image"],
            "properties": {
                "_id": {
                    "type": "string",
                    "uniqueItems": true
                },
                "comment": {
                    "type": "string",
                    "uniqueItems": true
                },
                "createdAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/User",
                    "uniqueItems": true
                },
                "image": {
                    "$ref": "#/definitions/Image",
                    "uniqueItems": true
                }
            }
        },
        "Image": {
            "required": ["_id", "imgUrl", "user", "path"],
            "properties": {
                "_id": {
                    "type": "string",
                    "uniqueItems": true
                },
                "destionation": {
                    "type": "string"
                },
                "encoding": {
                    "type": "string"
                },
                "fieldname": {
                    "type": "string"
                },
                "mimetype": {
                    "type": "string"
                },
                "originalname": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": "number",
                "imgUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/User"
                },
                "likes": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "likers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User"
                    }
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Comment"
                    }
                }
            }
        },
...
    }
}